+++ Movie/src/main/java/com/bank/interview/movie/service/dto/MovieRequestDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class MovieRequestDto {
 
     private String title;
     private String category;
-    private Float rating;
+    private BigDecimal rating;
 
 }
